Update htmldeps to be in three files. Remove X htmldeps; Fix textarea autoresize to resize on load.
