+++ shiny/experimental/ui/_accordion.py
@@ -10,7 +10,7 @@ from ..._namespaces import resolve_id_or_none
 from ..._utils import drop_none
 from ...session import require_active_session
 from ...types import MISSING, MISSING_TYPE
-from ...ui._html_deps_external import accordion_dependency
+from ...ui._html_deps_shinyverse import accordion_dependency
 from ...ui._tag import consolidate_attrs
 from ...ui.css_unit import CssUnit, as_css_unit
