+++ shiny/ui/_input_slider.py
@@ -16,7 +16,7 @@ from htmltools import HTML, Tag, TagAttrValue, TagChild, css, div, tags
 from .._docstring import add_example
 from .._namespaces import resolve_id
 from .._typing_extensions import NotRequired, TypedDict
-from ._html_deps import ionrangeslider_deps
+from ._html_deps_external import ionrangeslider_deps
 from ._utils import shiny_input_label
 
 # TODO: validate value(s) are within (min,max)?
