+++ shiny/ui/_layout.py
@@ -4,7 +4,7 @@ from typing import Literal, Optional
 
 from htmltools import Tag, TagAttrs, TagAttrValue, TagChild, css, div
 
-from ._html_deps_external import grid_dependency
+from ._html_deps_shinyverse import grid_dependency
 from ._tag import consolidate_attrs
 from ._utils import is_01_scalar
 from .css_unit import CssUnit, as_css_unit
