+++ shiny/ui/_web_component.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from htmltools import Tag, TagAttrs, TagAttrValue, TagChild
 
-from ._html_deps_external import web_component_dependency
+from ._html_deps_shinyverse import web_component_dependency
 
 
 def web_component(
