--- shiny/www/shared/_x/bslib/components/page_sidebar/page_sidebar.css
@@ -1,1 +1,0 @@
-.bslib-page-title{background-color:#212529;color:#fff;font-size:1.5rem;font-weight:300;padding:var(--bslib-spacer, 1rem);padding-left:1.5rem;margin-bottom:0}
