+++ shiny/www/shared/bslib/components/accordion/accordion.css
