+++ shiny/www/shared/bslib/components/sidebar/sidebar.css
