+++ shiny/www/shared/py-shiny/text-area/textarea-autoresize.css
@@ -1,0 +1,7 @@
+/* text-area/textarea-autoresize.css */
+.textarea-autoresize textarea.form-control {
+  padding: 5px 8px;
+  resize: none;
+  overflow-y: hidden;
+  height: auto;
+}
